src/lib/functions/run_pca.py
try:
    from sklearn.decomposition import PCA
    from sklearn.preprocessing import StandardScaler
except ImportError:
    print("Error in run pca function make sure the sklearn librarie is installed.")
    exit()

def run_pca(data):
    scaler = StandardScaler()
    scaled_data = scaler.fit_transform(data)
    pca = PCA(n_components=3)
    pca_result = pca.fit_transform(scaled_data)
    return pca_result