src/requirements.txt
streamlit==1.12.0
pandas==2.2.2
numpy==2.1.1
scikit-learn==1.1.1
umap-learn==0.5.3
plotly==5.9.0
openpyxl==3.0.10
# this versions are last modified in 7-9-24